email for paid contract
